+++ src/compas_ghpython/artists/robotmodelartist.py
@@ -48,4 +48,4 @@ class RobotModelArtist(BaseRobotModelArtist, BaseArtist):
 
     @staticmethod
     def draw_collection(collection):
-        pass
+        raise NotImplementedError
